+++ SQLiteStudio3/coreSQLiteStudio/services/functionmanager.h
@@ -31,6 +31,7 @@ class API_EXPORT FunctionManager : public QObject
             virtual QString toString() const;
 
             static QString typeString(Type type);
+            static QString displayString(Type type);
             static Type typeString(const QString& type);
 
             QString name;