+++ SQLiteStudio3/guiSQLiteStudio/windows/codesnippeteditormodel.h
@@ -38,6 +38,7 @@ class GUI_API_EXPORT CodeSnippetEditorModel : public QAbstractListModel
         QModelIndex moveUp(const QModelIndex& idx);
         QModelIndex moveDown(const QModelIndex& idx);
         int rowCount(const QModelIndex& parent = QModelIndex()) const override;
+        int columnCount(const QModelIndex& parent = QModelIndex()) const override;
         QVariant data(const QModelIndex& index, int role) const override;
         bool setData(const QModelIndex& index, const QVariant& value, int role) override;
         Qt::ItemFlags flags(const QModelIndex &index) const override;